Attempting to make "2022" fit inside the "years" box
This is only an attempt. I'm changing the size of the column "height" but i don't know how could I visualize the box after this change. Therefore, I don't know if this change works. (The choice of height = 300 was arbitrary. I just want to test any height larger than 260 and see if it works)