Creating the English READ.ME file
I asked ChatGPT to translate the whole READ.ME file of this repository, I still have to check for syntax errors and other possible problems, but I think it did a rather good job.
I had to divide the READ.ME in 4 parts and feed each one of them independently to ChatGPT (I'm using GPT 3.5 at the moment), since the whole file was way too big for a single prompt.